- Split SqlAgentTest for each method. - Fixed SqlQuery Javadoc comment. - Refactoring SqlQuery internal processing. - Add SqlEntityQuery#one and testcase.
fixed #201
