+++ src/test/java/jp/co/future/uroborosql/SqlUpdateTest.java
@@ -1,0 +1,88 @@
+package jp.co.future.uroborosql;
+
+import static org.junit.Assert.*;
+
+import java.math.BigDecimal;
+import java.nio.file.Paths;
+import java.sql.JDBCType;
+import java.sql.Types;
+import java.util.Arrays;
+import java.util.List;
+import java.util.Map;
+import java.util.stream.Collectors;
+
+import org.junit.Test;
+
+import jp.co.future.uroborosql.context.SqlContext;
+import jp.co.future.uroborosql.converter.MapResultSetConverter;
+import jp.co.future.uroborosql.exception.UroborosqlRuntimeException;
+import jp.co.future.uroborosql.utils.CaseFormat;
+
+public class SqlUpdateTest extends AbstractDbTest {
+	/**
+	 * DB更新処理のテストケース。
+	 */
+	@Test
+	public void testExecuteUpdate() throws Exception {
+		// 事前条件
+		cleanInsert(Paths.get("src/test/resources/data/setup", "testExecuteUpdate.ltsv"));
+
+		SqlContext ctx = agent.contextFrom("example/selectinsert_product")
+				.param("product_id", new BigDecimal("0"), JDBCType.DECIMAL)
+				.param("jan_code", "1234567890123", Types.CHAR);
+
+		int updateCount = agent.update(ctx);
+		assertEquals("データの登録に失敗しました。", 1, updateCount);
+
+		// 検証処理
+		List<Map<String, Object>> expectedDataList = getDataFromFile(Paths.get(
+				"src/test/resources/data/expected/SqlAgent", "testExecuteUpdate.ltsv"));
+		List<Map<String, Object>> actualDataList = agent.query("example/select_product")
+				.param("product_id", Arrays.asList(0, 1))
+				.stream(new MapResultSetConverter(agent.getSqlConfig().getDialect(), CaseFormat.LOWER_SNAKE_CASE))
+				.collect(Collectors.toList());
+
+		assertEquals(expectedDataList.toString(), actualDataList.toString());
+	}
+
+	/**
+	 * DB更新処理のテストケース。(Fluent API)
+	 */
+	@Test
+	public void testUpdateFluent() throws Exception {
+		// 事前条件
+		cleanInsert(Paths.get("src/test/resources/data/setup", "testExecuteUpdate.ltsv"));
+
+		int updateCount = agent.update("example/selectinsert_product")
+				.param("product_id", new BigDecimal("0"), JDBCType.DECIMAL)
+				.param("jan_code", "1234567890123", Types.CHAR).count();
+		assertEquals("データの登録に失敗しました。", 1, updateCount);
+
+		// 検証処理
+		List<Map<String, Object>> expectedDataList = getDataFromFile(Paths.get(
+				"src/test/resources/data/expected/SqlAgent", "testExecuteUpdate.ltsv"));
+		List<Map<String, Object>> actualDataList = agent.query("example/select_product")
+				.param("product_id", Arrays.asList(0, 1))
+				.stream(new MapResultSetConverter(agent.getSqlConfig().getDialect(), CaseFormat.LOWER_SNAKE_CASE))
+				.collect(Collectors.toList());
+
+		assertEquals(expectedDataList.toString(), actualDataList.toString());
+	}
+
+	/**
+	 * SQLファイルが存在しない場合のテストケース。
+	 */
+	@Test
+	public void testNotFoundFile() throws Exception {
+		try {
+			SqlContext ctx = agent.contextFrom("file");
+			agent.update(ctx);
+			// 例外が発生しなかった場合
+			fail();
+		} catch (UroborosqlRuntimeException ex) {
+			// OK
+		} catch (Exception e) {
+			fail(e.getMessage());
+		}
+	}
+}
